kern_shutdown: Always request a core dump in `vpanic()`
... even for a panic triggered while handling another one.

[Why]
If some code crashes while a panic is handled, we didn't get a kernel
core dump so far.

It is frustrating while working on the DRM drivers and their integration
with vt(4) because it makes debugging more difficult. Indeed, a panic
will usually display something on the console. If displaying something
on the console triggered another panic in vt(4) or the DRM drivers
before this patch, we wouldn't get any core dump.

[How]
The `RB_DUMP` flag is always set, not only for the top-level panic.